+++ objectbox/CHANGELOG.md
@@ -18,6 +18,9 @@
   `UniqueViolationException` if there is more than one result.
 * Update: [objectbox-c 0.18.0](https://github.com/objectbox/objectbox-c/releases/tag/v0.18.0).
 * Update: [objectbox-android 3.5.0](https://github.com/objectbox/objectbox-java/releases/tag/V3.5.0).
+* Update: [objectbox-swift 1.8.0](https://github.com/objectbox/objectbox-swift/releases/tag/v1.8.0).
+  Note that for this release `sync_flutter_libs` does not support macOS and iOS, please stick to
+  ObjectBox Dart 1.6.2.
 
 ## 1.6.2 (2022-08-24)